requirements.txt
streamlit
google-cloud-texttospeech
pydub
PyMuPDF
python-docx
ffmpeg-python